requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
asyncpg==0.29.0
psycopg2-binary==2.9.9
aioodbc==0.2.0
pyodbc==5.0.1
